requirements.txt
discord.py
python-dotenv
pronotepy
